refactor(knowledge): split lib.rs into responsibility submodules
The 979-line `primer-knowledge/src/lib.rs` becomes a directory-style
module split by responsibility. Behaviour-preserving: the 14-symbol
external `pub` surface is byte-identical, and all 25 crate tests pass
(baseline match).

- lib.rs (295): SqliteKnowledgeBase struct + lifecycle/diagnostic
  accessors + the BM25-only `impl KnowledgeBase`; module decls +
  `pub use schema::USER_VERSION` facade.
- ingest.rs (141): passage write path (plain / embedded / --reembed).
- retrieval.rs (155): hybrid `retrieve_hybrid` + `knn_scan` read path.
- schema.rs (226): table creation + migrations + `pub USER_VERSION`.
- locale_sql.rs (78): `LocaleSql` (now `pub(crate)`).
- vector.rs (46), embedding_model.rs (84), fts.rs (37): pure leaf helpers.

Cross-module helpers bumped to `pub(crate)` (invisible to the external
pub-surface diff). tests.rs gains explicit imports for the helpers the
`use super::*` facade no longer carries (`table_exists`, `vec_to_blob`,
`blob_to_vec`, and the `Embedder` trait).

Verification: primer-knowledge 25/25 (baseline match), full-workspace
`cargo test` green (51 ok, 0 failed), clippy -D warnings clean, fmt clean,
pub-surface diff empty. CLAUDE.md updated with the module layout;
README/ROADMAP untouched (internal refactor).
